curl: Group curl related functions in a new header file
Also rename some files to a proper name.
