+++ src/swupd-curl-internal.h
